Implemented Boost.Chrono to determine algorithm speed Fixed issue #2 (TEMP) (Input limit set) Fixed issue #3 Return type changed to (u)long long to fast int type for optimizations. optimized imports cosmetic cleanup

+++ src/LargestPrime.h
@@ -1,4 +1,7 @@
 #ifndef _GREATESTPRIME_LARGESTPRIME_H_
 #define _GREATESTPRIME_LARGESTPRIME_H_
-unsigned long long largestPrime(unsigned long long upperBound);
+
+#include <stdint.h>
+
+int_fast64_t largestPrime(int_fast64_t upperBound);
 #endif //_GREATESTPRIME_LARGESTPRIME_H_
